DRC-2559: Allow variable clients to specify data frequency. does not fixbamboo

+++ RobotDataCommunication/src/us/ihmc/robotDataCommunication/YoVariablesUpdatedListener.java
@@ -28,7 +28,7 @@ public interface YoVariablesUpdatedListener
 
    public void receiveTimedOut();
 
-   public long getDisplayOneInNPackets();
+   public int getDisplayOneInNPackets();
 
    public void timestampReceived(long timestamp);
